+++ app/src/main/java/com/gowthamraj07/journeytracker/domain/repository/PlacesRepository.kt
@@ -5,5 +5,5 @@ import kotlinx.coroutines.flow.Flow
 
 
 interface PlacesRepository {
-    suspend fun loadPlacesFor(tripId: Int): Flow<List<Place>>
+    suspend fun loadPlacesFor(tripId: Long): Flow<List<Place>>
 }
