+++ app/src/main/java/com/gowthamraj07/journeytracker/domain/usecase/LoadPlacesUseCase.kt
@@ -5,7 +5,7 @@ import com.gowthamraj07.journeytracker.domain.repository.PlacesRepository
 import kotlinx.coroutines.flow.Flow
 
 class LoadPlacesUseCase(private val placesRepository: PlacesRepository) {
-    suspend fun execute(tripId: Int): Flow<List<Place>> {
+    suspend fun execute(tripId: Long): Flow<List<Place>> {
         return placesRepository.loadPlacesFor(tripId)
     }
 }
